+++ MainServer/src/main/java/com/pxl/emotionjava/services/api/ProfileDataService.java
@@ -5,13 +5,13 @@ import com.pxl.emotionjava.entities.Profile;
 import java.util.List;
 
 public interface ProfileDataService {
-    Profile getProfileById(int id);
+    Profile getProfileById(Long id);
 
-    List<Profile> getProfilesByUserId(int id);
+    List<Profile> getProfilesByUserId(Long id);
 
     String addOrUpdateProfile(Profile profile);
 
-    String deleteProfile(int id);
+    String deleteProfile(Long id);
 
     List<Profile> getAllProfiles();
 }
